created simple scene
character can't move due to Object reference not set to an instance of an object. The code needs to be debugged and fixed in order for the character to move. Action map is created and attached to the character. No other feature were added because I was struggling too much on fixing the broken code.